+++ src/thread.h
@@ -76,7 +76,7 @@ void mainthread_search();
 
 struct ThreadPool {
   Pos *pos[MAX_THREADS];
-  size_t num_threads;
+  int num_threads;
 #ifndef __WIN32__
   pthread_mutex_t mutex;
   pthread_cond_t sleepCondition;
@@ -91,7 +91,7 @@ typedef struct ThreadPool ThreadPool;
 void threads_init(void);
 void threads_exit(void);
 void threads_start_thinking(Pos *pos, LimitsType *);
-void threads_set_number(size_t num);
+void threads_set_number(int num);
 uint64_t threads_nodes_searched(void);
 uint64_t threads_tb_hits(void);
